refactor: header 컴포넌트 구조 개선
